refactor: 트레이너 스케줄 리팩토링 - 코드 개선, 휴무일 변경 확인 로직 추가 등
