+++ src/feature/schedule/ui/WeeklyTimetable/ScheduleItemBottomSheet.tsx
@@ -25,7 +25,7 @@ import { useTrainerChangeReservationMutation } from '../../api/useTrainerChangeR
 import { useTrainerChangeShowMutation } from '../../api/useTrainerChangeShowMutation';
 import { FlatSchedule } from '../../model/type';
 
-const BottomSheet = ({ schedule }: { schedule: FlatSchedule }) => {
+const ScheduleItemBottomSheet = ({ schedule }: { schedule: FlatSchedule }) => {
   return sheetMapper(schedule)[schedule.reservationStatus];
 };
 
@@ -106,10 +106,10 @@ const CompletedSheet = ({ schedule }: { schedule: FlatSchedule }) => {
   }
 
   return (
-    <SheetContent side='bottom'>
+    <SheetContent side='bottom' className='p-7 pb-11'>
       <div className='flex w-full flex-col'>
         <h3 className={cn(Typography.HEADING_3)}>{applicantName}</h3>
-        <p className={cn(Typography.BODY_1, 'mt-2')}>{title}</p>
+        <p className={cn(Typography.BODY_1, 'mt-2')}>{title}123</p>
         <div className={cn('mt-7 flex w-full justify-center gap-6')}>
           {!isBefore && (
             <button
@@ -181,7 +181,7 @@ const AvailableSheet = ({ schedule }: { schedule: FlatSchedule }) => {
   }
 
   return (
-    <SheetContent side='bottom'>
+    <SheetContent side='bottom' className='p-7 pb-11'>
       <div className='flex w-full flex-col'>
         <h3 className={cn(Typography.HEADING_3)}>{title}</h3>
         <div className={cn('mt-7 flex w-full justify-center gap-6')}>
@@ -234,7 +234,7 @@ const NoShowSheet = ({ schedule }: { schedule: FlatSchedule }) => {
   };
 
   return (
-    <SheetContent side='bottom'>
+    <SheetContent side='bottom' className='p-7 pb-11'>
       <div className='flex w-full flex-col'>
         <h3 className={cn(Typography.HEADING_3)}>
           {applicantName}
@@ -304,7 +304,7 @@ const DisabledSheet = ({ schedule }: { schedule: FlatSchedule }) => {
   }
 
   return (
-    <SheetContent side='bottom'>
+    <SheetContent side='bottom' className='p-7 pb-11'>
       <div className='flex w-full flex-col'>
         <h3 className={cn(Typography.HEADING_3)}>{title}</h3>
         <p className={cn(Typography.BODY_1, 'mt-2 text-point')}>예약불가</p>
@@ -324,4 +324,4 @@ const DisabledSheet = ({ schedule }: { schedule: FlatSchedule }) => {
   );
 };
 
-export { BottomSheet };
+export { ScheduleItemBottomSheet };
